Move NodeSvcHandle into core networking
